+++ src/main/java/techcourse/myblog/config/NoneAuthInterceptor.java
@@ -1,4 +1,4 @@
-package techcourse.myblog.web;
+package techcourse.myblog.config;
 
 import org.springframework.web.servlet.HandlerInterceptor;
 
@@ -6,16 +6,17 @@ import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import javax.servlet.http.HttpSession;
 
-public class NeedAuthInterceptor implements HandlerInterceptor {
+public class NoneAuthInterceptor implements HandlerInterceptor {
 
     @Override
     public boolean preHandle(final HttpServletRequest req, final HttpServletResponse res, final Object handler) throws Exception {
         HttpSession session = req.getSession();
-        if (session.getAttribute("username") != null) {
-            res.sendRedirect("/");
+
+        if (session.getAttribute("username") == null) {
+            res.sendRedirect("/auth/login");
             return false;
         }
+
         return true;
     }
-
 }
