+++ src/main/java/techcourse/myblog/config/WebConfig.java
@@ -3,8 +3,6 @@ package techcourse.myblog.config;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.web.servlet.config.annotation.InterceptorRegistry;
 import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
-import techcourse.myblog.web.NeedAuthInterceptor;
-import techcourse.myblog.web.NoneAuthInterceptor;
 
 @Configuration
 public class WebConfig implements WebMvcConfigurer {
